requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==1.10.15
python-multipart==0.0.9
python-dotenv==1.0.1
supabase==2.6.0
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
